requirements.txt
# ============================================================
#  Robin v4.0 - Requirements
#  Installieren mit:  pip install -r requirements.txt --break-system-packages
# ============================================================

# --- Kern (bereits v3) ---
openai-whisper>=20231117
faster-whisper>=1.0.0
piper-tts>=1.2.0
pyttsx3>=2.90
sounddevice>=0.4.6
numpy>=1.24.0
ollama>=0.4.0
psutil>=5.9.0
pyautogui>=0.9.54
screen-brightness-control>=0.22.0
pycaw>=20230407
comtypes>=1.4.0

# --- PC-Steuerung ---
pyperclip>=1.8.2
pygetwindow>=0.0.9
pywin32>=306

# --- Medien ---
spotipy>=2.23.0           # Spotify API (optional, nur wenn CLIENT_ID konfiguriert)
Pillow>=10.0.0            # Screenshot-Analyse

# --- Produktivitaet ---
wikipedia>=1.4.0           # Wikipedia-Abfragen
requests>=2.31.0           # Waehrungsumrechnung, APIs
google-api-python-client>=2.100.0   # Google Calendar (optional)
google-auth-oauthlib>=1.1.0         # Google OAuth (optional)

# --- Dokument-Zusammenfassung ---
pdfplumber>=0.10.0        # PDF-Lesen
python-docx>=1.1.0        # DOCX-Lesen

# --- Kommunikation ---
pywhatkit>=5.4            # WhatsApp Web (optional)

# --- System-Monitoring ---
GPUtil>=1.4.0              # GPU-Monitoring (optional, fuer NVIDIA)

# --- Audio (Lautstaerkesteuerung bei TTS) ---
scipy>=1.11.0              # fuer WAV-Verarbeitung (optional)
